Add support for base24 and rename to tinted-gallery

+++ build.sh
@@ -1,8 +1,22 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 set -eu pipefail
 
+main() {
+  setup
+
+  generate_themes "base16"
+  generate_themes "base24"
+
+  erb template.erb > out/index.html
+}
+
 setup() {
+  if [ -z "$(command -v 'vim')" ]; then
+    echo "Error: vim is required for build."
+    exit 1
+  fi
+
   rm -rf out
   mkdir out
 
@@ -13,17 +27,17 @@ setup() {
   git clone --depth=1 https://github.com/tinted-theming/schemes
 }
 
-main() {
-  setup
+generate_themes() {
+  local system=${1:-"base16"}
 
-  colorschemes=($(ls schemes/base16/ | grep yaml | sed 's/\..*$//'))
+  colorschemes=($(ls "schemes/$system/" | grep yaml | sed 's/\..*$//'))
   tmpfile="$0.html"
 
   vim -es -u NORC -N \
     -c 'silent! set termguicolors' \
-    -c 'silent! set runtimepath+=base16-vim' \
+    -c 'silent! set runtimepath+=tinted-vim' \
     -c 'silent! syntax on' \
-    -c "silent! colorscheme base16-${colorschemes[0]}" \
+    -c "silent! colorscheme $system-${colorschemes[0]}" \
     -c 'silent! TOhtml' \
     -c 'silent! wqa!' \
     $0 > /dev/null 2>&1
@@ -36,19 +50,17 @@ main() {
       -c 'set termguicolors' \
       -c 'set runtimepath+=tinted-vim' \
       -c 'syntax on' \
-      -c "colorscheme base16-$colorscheme" \
+      -c "colorscheme $system-$colorscheme" \
       -c 'TOhtml' \
       -c 'wqall' \
       $0 > /dev/null 2>&1
 
     grep -Pzo '(?s)<style>.*</style>' $tmpfile \
-      | sed "3,14!d;s/body/pre/;s/^/#base16-$colorscheme /" \
-      > out/$colorscheme.css
+      | sed "3,14!d;s/body/pre/;s/^/#$system-$colorscheme /" \
+      > "out/$system-$colorscheme.css"
 
     rm -f $tmpfile
   done
-
-  erb template.erb > out/index.html
 }
 
 main
